+++ src/main/java/de/MCmoderSD/utilities/other/Calculate.java
@@ -220,7 +220,7 @@ public class Calculate {
         return "[" + new java.text.SimpleDateFormat("dd-MM-yyyy|HH:mm:ss").format(getTimestamp()) + "]";
     }
 
-    public static boolean checkAge(byte minAge, Birthdate birthdate) {
-        return (byte) (Calendar.getInstance(birthdate.getTimeZone()).get(Calendar.YEAR) - birthdate.getYear()) >= minAge;
+    public static boolean checkAge(int minAge, Birthdate birthdate) {
+        return (Calendar.getInstance(birthdate.getTimeZone()).get(Calendar.YEAR) - birthdate.getYear()) >= minAge;
     }
 }
